+++ Assets/Examples/DefaultNodes/Editor/IfNodeView.cs
@@ -12,6 +12,7 @@ public class IfNodeView : BaseNodeView
 {
 	public override void Enable()
 	{
+		hasSettings = true;	// or base.Enable();
 		var node = nodeTarget as IfNode;
 
         // Create your fields using node's variables and add them to the controlsContainer